Move PD queries to new module
