+++ openlibrary/plugins/upstream/account.py
@@ -1,7 +1,6 @@
 import json
 import logging
 from collections.abc import Callable, Iterable, Mapping
-from dataclasses import dataclass
 from datetime import datetime
 from math import ceil
 from typing import TYPE_CHECKING, Any, Final
@@ -30,7 +29,7 @@ from openlibrary.accounts import (
     valid_email,
 )
 from openlibrary.accounts.model import sendmail
-from openlibrary.core import cache, lending, stats
+from openlibrary.core import lending, stats
 from openlibrary.core import helpers as h
 from openlibrary.core.booknotes import Booknotes
 from openlibrary.core.bookshelves import Bookshelves
@@ -43,10 +42,11 @@ from openlibrary.core.observations import Observations
 from openlibrary.core.ratings import Ratings
 from openlibrary.i18n import gettext as _
 from openlibrary.plugins import openlibrary as olib
+from openlibrary.plugins.openlibrary.pd import get_pd_options, get_pd_org
 from openlibrary.plugins.recaptcha import recaptcha
 from openlibrary.plugins.upstream import borrow, forms, utils
 from openlibrary.plugins.upstream.mybooks import MyBooksTemplate
-from openlibrary.utils.dateutil import DAY_SECS, elapsed_time
+from openlibrary.utils.dateutil import elapsed_time
 
 if TYPE_CHECKING:
     from openlibrary.plugins.upstream.models import Work
@@ -269,55 +269,6 @@ class account(delegate.page):
     def GET(self):
         user = accounts.get_current_user()
         return render.account(user)
-
-
-@dataclass
-class PDOption:
-    """Represents an option in the print-disability qualifying organization selector"""
-
-    label: str
-    value: str
-
-
-def get_pd_org(identifier):
-    orgs = cached_pd_org_query()
-    return next((org for org in orgs if org['identifier'] == identifier), None)
-
-
-def get_pd_options():
-    options = [
-        PDOption("BARD", "ia_nlsbardaccess_disabilityresources"),
-        PDOption("BookShare", "ia_bookshareaccess_disabilityresources"),
-        PDOption("ACE", "aceportalocul_disabilityresources"),
-        PDOption("I don't have one yet", "unqualified"),
-    ]
-
-    pd_orgs = cached_pd_org_query()
-    options += [PDOption(org.get("title"), org.get("identifier")) for org in pd_orgs]
-
-    return options
-
-
-def make_pd_org_query():
-    base_url = config.get("bookreader_host", "")
-    if not base_url:
-        return []
-    params = "q=collection:print_disability_access&fl[]=identifier,title&rows=1000&page=1&output=json"
-    url = f"https://{base_url}/advancedsearch.php?{params}"
-    try:
-        response = requests.get(url)
-        response.raise_for_status()
-    except requests.HTTPError:
-        return []
-    except requests.exceptions.JSONDecodeError:
-        return []
-
-    return response.json().get("response", {}).get("docs", []) or []
-
-
-@cache.memoize(engine="memcache", key="pd-org-query", expires=DAY_SECS)
-def cached_pd_org_query():
-    return make_pd_org_query()
 
 
 class account_create(delegate.page):
